RAG/RAG-Level-01/P01-RAG-Projects/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.3.0",
    "chromadb>=1.3.4",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.121.0",
    "jupyter>=1.1.1",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-groq>=1.0.0",
    "langchain-huggingface>=1.0.1",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.2",
    "langserve>=0.3.3",
    "notebook>=7.4.7",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "sse-starlette>=3.0.3",
    "streamlit>=1.51.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.38.0",
    "wikipedia>=1.4.0",
]
[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
